requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
openai>=1.12.0
python-dotenv>=1.0.0
requests>=2.31.0